+++ Classes/GiocoBuilder.cs
@@ -8,8 +8,8 @@ namespace Classes
     {
         private int NumGiocatori { get; set; }
         private int NumMazziIniziali { get; set; }
-        private int PuntataMinima { get; set; }
-        private int? PuntataMassima { get; set; }
+        private double PuntataMinima { get; set; }
+        private double? PuntataMassima { get; set; }
         private bool Mischia { get; set; } = true;
         private int? RandomMischiata { get; set; }
         private int? PercMischiata { get; set; }
